+++ RandomEncounters/Encounter/EncounterEntry.cs
@@ -1,0 +1,9 @@
+﻿namespace RandomEncounters
+{
+    internal class EncounterEntry
+    {
+        public string Name;
+        public int Weight;
+        public System.Action Trigger;
+    }
+}